Extract JAR plugin loader into a separate class
